fix: add get-network-stats
