+++ applications/tari_base_node/src/commands/performer.rs
@@ -72,7 +72,6 @@ impl Performer {
             SearchUtxo => self.process_search_utxo(typed_args).await,
             SearchKernel => self.process_search_kernel(typed_args).await,
             Whoami => self.command_handler.whoami(),
-            GetNetworkStats => self.command_handler.get_network_stats(),
             Exit | Quit => {
                 println!("Shutting down...");
                 info!(
@@ -162,9 +161,6 @@ impl Performer {
                     "Display identity information about this node, including: public key, node ID and the public \
                      address"
                 );
-            },
-            GetNetworkStats => {
-                println!("Displays network stats");
             },
             Exit | Quit => {
                 println!("Exits the base node");
